requirements.txt
customtkinter
requests
beautifulsoup4
pillow
py7zr
pyinstaller
